requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.9
django-3-jet==1.0.8
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
importlib-metadata==4.8.2
Markdown==3.3.6
Pillow==8.4.0
pycodestyle==2.8.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uuid==1.30
zipp==3.6.0
